Remove IndividualTrait.php, use common module for common stuff
