--- .github/workflows/phpunit.yml
@@ -1,29 +1,0 @@
-name: PHPUnit
-
-on: [push, pull_request]
-
-jobs:
-  run:
-    runs-on: ${{ matrix.operating-system }}
-
-    strategy:
-      matrix:
-        operating-system: [ubuntu-22.04]
-        php-version: ['8.1', '8.2']
-
-    name: Testing PHP ${{ matrix.php-version }} on ${{ matrix.operating-system }}
-
-    steps:
-      - uses: shivammathur/setup-php@master
-        with:
-          php-version: ${{ matrix.php-version }}
-          extensions: json,intl,mbstring
-          coverage: pcov
-
-      - name: Checkout repository
-        uses: actions/checkout@v3
-
-      - run: composer validate
-      - run: composer install --no-progress
-      - run: composer require phpdocumentor/reflection-docblock
-      - run: vendor/bin/phpunit --coverage-clover=tests/coverage.xml
