+++ composer.json
@@ -22,6 +22,8 @@
             "magicsunday/webtrees-module-installer-plugin": true
         }
     },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "require": {
         "ext-json": "*",
         "ext-dom": "*",
@@ -30,7 +32,6 @@
         "magicsunday/webtrees-module-installer-plugin": "^1.3"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.5",
         "phpstan/phpstan": "^1.9",
         "phpstan/phpstan-strict-rules": "^1.4",
         "phpstan/phpstan-deprecation-rules": "^1.1",
